rag-app/requirements.txt
streamlit
chroma
google-genai
langchain
pypdf
python-docx
python-dotenv
